issue #18 (#314) Curve mode reverse mode should reverse playback
