+++ src/apps/sequencer/engine/CurveTrackEngine.cpp
@@ -14,10 +14,50 @@
 
 static Random rng;
 
+static const std::map<int, Curve::Type> REV_SHAPE_MAP = {
+        {0, Curve::Low},
+        {1, Curve::High},
+        {2, Curve::RampDown},
+        {3, Curve::RampUp},
+        {4, Curve::rampDownHalf},
+        {5, Curve::rampUpHalf},
+        {6, Curve::doubleRampDownHalf},
+        {7, Curve::doubleRampUpHalf},
+        {8, Curve::ExpDown},
+        {9, Curve::ExpUp},
+        {10, Curve::expDownHalf},
+        {11, Curve:: expUpHalf},
+        {12, Curve::doubleExpDownHalf},
+        {13, Curve::doubleExpUpHalf},
+        {14, Curve::LogDown},
+        {15, Curve::LogUp},
+        {16, Curve::logDownHalf},
+        {17, Curve::logUpHalf},
+        {18, Curve::doubleLogDownHalf},
+        {19, Curve::doubleLogUpHalf},
+        {20, Curve::SmoothDown},
+        {21, Curve::SmoothUp},
+        {22, Curve::smoothDownHalf},
+        {23, Curve::smoothUpHalf},
+        {24, Curve::doubleSmoothDownHalf},
+        {25, Curve::doubleSmoothUpHalf},
+        {26, Curve::Triangle},
+        {27, Curve::RevTriangle},
+        {28, Curve::Bell},
+        {29, Curve::RevBell},
+        {30, Curve::StepDown},
+        {31, Curve::StepUp},
+        {32, Curve::ExpUp2x},
+        {33, Curve::ExpDown2x},
+        {34, Curve::ExpUp3x},
+        {35, Curve::ExpUp4x},
+        {36, Curve::ExpDown4x}
+    };
+
 static float evalStepShape(const CurveSequence::Step &step, bool variation, bool invert, float fraction, int direction) {
     auto function = Curve::function(Curve::Type(variation ? step.shapeVariation() : step.shape()));
     if (direction == -1) {
-        function = Curve::function(Curve::Type(variation ? step.shapeVariation() : Curve::revAt(step.shape())));
+        function = Curve::function(Curve::Type(variation ? step.shapeVariation() : REV_SHAPE_MAP.at(step.shape())));
     }
     float value = function(fraction);
     if (invert) {
