+++ src/apps/sequencer/ui/pages/CurveSequenceEditPage.cpp
@@ -38,6 +38,46 @@ enum class Function {
 
 static const char *functionNames[] = { "SHAPE", "MIN", "MAX", "GATE", nullptr };
 
+static const std::map<int, Curve::Type> INV_SHAPE_MAP = {
+        {0, Curve::High},
+        {1, Curve::Low},
+        {2, Curve::RampDown},
+        {3, Curve::RampUp},
+        {4, Curve::rampDownHalf},
+        {5, Curve::rampUpHalf},
+        {6, Curve::doubleRampDownHalf},
+        {7, Curve::doubleRampUpHalf},
+        {8, Curve::ExpDown},
+        {9, Curve::ExpUp},
+        {10, Curve::expDownHalf},
+        {11, Curve:: expUpHalf},
+        {12, Curve::doubleExpDownHalf},
+        {13, Curve::doubleExpUpHalf},
+        {14, Curve::LogDown},
+        {15, Curve::LogUp},
+        {16, Curve::logDownHalf},
+        {17, Curve::logUpHalf},
+        {18, Curve::doubleLogDownHalf},
+        {19, Curve::doubleLogUpHalf},
+        {20, Curve::SmoothDown},
+        {21, Curve::SmoothUp},
+        {22, Curve::smoothDownHalf},
+        {23, Curve::smoothUpHalf},
+        {24, Curve::doubleSmoothDownHalf},
+        {25, Curve::doubleSmoothUpHalf},
+        {26, Curve::RevTriangle},
+        {27, Curve::Triangle},
+        {28, Curve::RevBell},
+        {29, Curve::Bell},
+        {30, Curve::StepDown},
+        {31, Curve::StepUp},
+        {32, Curve::ExpUp2x},
+        {33, Curve::ExpDown2x},
+        {34, Curve::ExpUp3x},
+        {35, Curve::ExpUp4x},
+        {36, Curve::ExpDown4x}
+    };
+
 
 static const CurveSequenceListModel::Item quickEditItems[8] = {
     CurveSequenceListModel::Item::FirstStep,
@@ -278,7 +318,7 @@ void CurveSequenceEditPage::draw(Canvas &canvas) {
     if (isActiveSequence) {
         canvas.setColor(Color::Bright);
 
-        int x;
+        int x = 0;
         auto dir = trackEngine.sequenceState().direction();
         if (dir == 1) {
             x = ((trackEngine.currentStep() - stepOffset) + trackEngine.currentStepFraction()) * stepWidth;
@@ -384,7 +424,7 @@ void CurveSequenceEditPage::keyPress(KeyPressEvent &event) {
         for (size_t stepIndex = 0; stepIndex < _stepSelection.size(); ++stepIndex) {
         if (_stepSelection[stepIndex]) {
             auto &step = sequence.step(stepIndex);
-            auto reverseShape = Curve::invAt(step.shape());
+            auto reverseShape = INV_SHAPE_MAP.at(step.shape());
             step.setShape(reverseShape);
         }
     }
